Rename spec/unit/puppet_pal_2pec.rb to spec/unit/puppet_pal_spec.rb
